Clean up example and mimic AudioTesting.
