+++ packages/example/src/Video.tsx
@@ -288,7 +288,7 @@ export const Index: React.FC = () => {
 				width={1080}
 				height={1080}
 				fps={30}
-				durationInFrames={150}
+				durationInFrames={300}
 			/>
 			<Still id="Orb" component={OrbScene} width={2000} height={2000} />
 		</>
